Move new defn api to Schema namespace
