+++ lib/graphql.rb
@@ -117,11 +117,4 @@ require "graphql/subscriptions"
 require "graphql/parse_error"
 require "graphql/backtrace"
 
-require "graphql/schema_member"
-require "graphql/object"
-require "graphql/enum"
-require "graphql/interface"
-require "graphql/union"
-require "graphql/input_object"
-require "graphql/scalar"
 require "graphql/deprecated_dsl"
