+++ lib/graphql/query.rb
@@ -240,7 +240,7 @@ module GraphQL
         value = abstract_type
         abstract_type = nil
       end
-      if value.is_a?(GraphQL::Object)
+      if value.is_a?(GraphQL::Schema::Object)
         value = value.object
       end
       @resolved_types_cache[abstract_type][value]
