+++ lib/graphql/relay/global_id_resolve.rb
@@ -7,7 +7,7 @@ module GraphQL
       end
 
       def call(obj, args, ctx)
-        if obj.is_a?(GraphQL::Object)
+        if obj.is_a?(GraphQL::Schema::Object)
           obj = obj.object
         end
         type = @type.respond_to?(:graphql_definition) ? @type.graphql_definition : @type
